add: controller of Tweet and Video

+++ src/controllers/tweet.controllers.js
@@ -36,15 +36,13 @@ const createTweet = asyncHandler(async (req, res) => {
                 "Error while creating tweet"
             )
         )
-    }
-    
-    const createdtweet = await Tweet.findById(tweetInstance._id) 
+    } 
 
 
     res.status(200)
         .json(
             new ApiResponse(
-                200 , createdtweet , "User is Created "
+                200 , tweetInstance , "User is Created "
             )
         )
 })
@@ -52,12 +50,13 @@ const createTweet = asyncHandler(async (req, res) => {
 const getUserTweets = asyncHandler(async (req, res) => {
     const userId = req.user?._id;
 
-    if ( !isValidObjectId(userId) ){
+    if (!userId && !isValidObjectId(userId) ){
         res.json(
             new ApiError( 408 , "Object Id is Invalid " )
         )
     }
 
+    /*
     const tweets = await User.aggregate([
         {
             $match: {
@@ -103,6 +102,74 @@ const getUserTweets = asyncHandler(async (req, res) => {
             }
         }
     ])
+        */
+
+
+    const tweets = await Tweet.aggregate([
+        {
+            $match: {
+                owner : new mongoose.Types.ObjectId(userId),
+            }
+        },
+        {
+            $lookup: {
+                from: "users",
+                localField: "owner",
+                foreignField: "_id",
+                as: "owner",
+                pipeline : [
+                    {
+                        $project : {
+                            username : 1,
+                            avatar : 1
+                        }
+                    }
+                ]
+            }
+        },
+        {
+            $lookup: {
+                from: "likes",
+                localField: "_id",
+                foreignField: "tweet",
+                as: "likes",
+            }
+        },
+        {
+            $addFields: {
+               likecount: {
+                $size : "$likes"
+               },
+               owner: {
+                $first : "$owner"
+               },
+               isLiked : {
+                   $cond : {
+                       if : {
+                           $in : [req?.user?._id , "$likes.likedBy"]
+                       },
+                       then : true,
+                       else : false
+                   }
+               }
+            }
+        },
+        {
+            $sort: {
+                createdAt : -1
+            }
+        },
+        {
+            $project: {
+                _id : 1,
+                content : 1,
+                createdAt : 1,
+                likecount : 1,
+                owner : 1,
+                isLiked : 1
+            }
+        }
+    ])
 
     if (!tweets) {
         res.json(
@@ -118,12 +185,24 @@ const getUserTweets = asyncHandler(async (req, res) => {
 })
 
 const updateTweet = asyncHandler(async (req, res) => {
-    const tweetId  = req.params?.tweetId;
+    const {tweetId}  = req.params?.tweetId;
 
-    if ( !tweetId) {
+    if ( !tweetId && !isValidObjectId(tweetId) ){ {
         res.json(
             new ApiError(409 , "Invalid Tweet Signature")
         )
+    }
+
+    const tweentInstance = await Tweet.findById(tweetId);
+
+    if ( !tweentInstance ){
+        res.json(
+            new ApiError(409 , "Invalid Tweet Signature")
+        )
+    }
+
+    if ( !tweentInstance?.owner.equals(req.user?._id) ){
+        throw  new ApiResponse(409 , "User is not authorized to update this tweet")
     }
 
     const { content } = req.body;
@@ -134,7 +213,7 @@ const updateTweet = asyncHandler(async (req, res) => {
         )
     }
 
-    const tweetInstance = await Tweet.findByIdAndUpdate(
+    const updatedtweetInstance = await Tweet.findByIdAndUpdate(
         tweetId,
         {
             $set : {
@@ -146,15 +225,16 @@ const updateTweet = asyncHandler(async (req, res) => {
         }
     )
 
-    if ( !tweetInstance ){
+    if ( !updatedtweetInstance ){
         res.json(
             new ApiError(407 , "Error in the Updatation of Tweet")
         )
     }
 
     res.json(
-        new ApiResponse(200 , tweetInstance , "Updation is Done Successfully !! ")
+        new ApiResponse(200 , updatedtweetInstance , "Updation is Done Successfully !! ")
     )
+}
 })
 
 const deleteTweet = asyncHandler(async (req, res) => {
@@ -166,15 +246,29 @@ const deleteTweet = asyncHandler(async (req, res) => {
         )
     }
 
+    const tweetInstance = await Tweet.findById(tweetId);
+
+    if ( !tweetInstance ){
+        throw new ApiError(409 , "Invalid Tweet Signature")
+    }
+
+    if (tweetInstance.owner.toString() !== req.user?._id.toString()) {
+        throw new ApiError(400, "Only tweet owner can delete this tweet");
+    }
+
+    
     const deleteTweetInstance = await Tweet.deleteOne({
         _id : new mongoose.Types.ObjectId(tweetId)
     })
-
+    
     if ( !deleteTweetInstance ){
         res.json(
-        new ApiError(408 , "Error While Deletion of the Tweet")    
+            new ApiError(408 , "Error While Deletion of the Tweet")    
         )
     }
+    // Likes and Comments deletion
+    await Like.deleteMany({ tweets: tweetId });
+    await Comment.deleteMany({ tweet: tweetId });
 
     res.json(
         new ApiResponse(200 , deleteTweetInstance , "Deletion Process is Done")
